+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 * Added the root effect of Binding Shot (Hunter) to Auras widget as CC aura [Comment #6416].
 * Fixed CC aura id for Clash (Brewmaster Monk) in Auras widget [Comment #6408].
+* Fixed a bug with threat values not being scaled correctly in TBC and Wrath Classic WoW [GH-346, Issue #521].